refactor(auth): SellerCredentialService 추출 + 판매자 흐름 분리 (P0-1 단계 6)
AuthService 의 판매자 자격증명 기반 흐름 4 종 (sellerLogin / refreshSeller /
logoutSeller / changeSellerPassword) 을 신규 SellerCredentialService 로 분리.
AuthController 가 직접 호출. AuthService 는 일반 유저 흐름 (refresh / logout /
issueDevAccessToken / me) 만 보유하게 됨.

- 신규: SellerCredentialService (interface + impl, SELLER_CREDENTIAL_SERVICE Symbol)
  - sellerLogin, refreshSeller, logoutSeller, changeSellerPassword
- 제거 (AuthService): 위 4 개 + 관련 의존 (ClockService / argon2 / SellerCredentialRepository
  / AuditLogRepository / Audit/Prisma enum / Account enum / AccountStatus 타입)
- AuthController: sellerAuth 의존 추가, /auth/seller/* 4 개 라우트가 sellerAuth.* 호출
- AuthModule: SellerCredentialService 등록
- 스펙 이전: auth.seller.service.spec.ts → services/seller-credential.service.spec.ts
  (대상이 AuthService → SellerCredentialService 로 변경, accounts mock 제거)
- AuthController spec: 판매자 라우트 mock 을 sellerAuth 로 분리

AuthService 줄 수: 약 290 → 약 80. logout / logoutSeller 책임 명확화 (각자
도메인 서비스에 거주, 90% 중복은 그대로 두되 더 이상 같은 클래스 안에서 혼재하지 않음).
